requirements.txt
discord.py
requests
asyncio
datetime
sly